+++ services/web-app/pages/assets/functions.js
@@ -1,0 +1,46 @@
+/*
+ * Copyright IBM Corp. 2021, 2021
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+import { NextSeo } from 'next-seo'
+import { useContext, useEffect } from 'react'
+
+import CatalogIndexPage from '@/components/catalog-index-page'
+import { assetsNavData } from '@/data/nav-data'
+import { LayoutContext } from '@/layouts/layout'
+import { getAllLibraries } from '@/lib/github'
+import styles from '@/pages/pages.module.scss'
+
+const Functions = ({ librariesData }) => {
+  const { setNavData } = useContext(LayoutContext)
+
+  const seo = {
+    title: 'Functions'
+  }
+
+  useEffect(() => {
+    setNavData(assetsNavData)
+  }, [setNavData])
+
+  return (
+    <>
+      <NextSeo {...seo} />
+      <CatalogIndexPage data={librariesData} type="function" />
+      <pre className={styles.data}>{JSON.stringify(librariesData, null, 2)}</pre>
+    </>
+  )
+}
+
+export const getStaticProps = async () => {
+  const librariesData = await getAllLibraries()
+
+  return {
+    props: {
+      librariesData
+    }
+  }
+}
+
+export default Functions